Delete Register 画面側作成2

+++ src/main/java/com/example/entity/BookDTO.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookDTO {
-    private int num;
+    private Long num;
     private String title;
     private int price;
     private String author;
